# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BlockVotes API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidate]
      summary: List candidates for an election
      parameters:
        - name: election_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidatesResponse"
    post:
      operationId: addCandidate
      tags: [candidate]
      summary: Add a candidate to an election
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCandidateInput"
      responses:
        "200":
          description: Candidate added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /candidates/{candidateId}:
    delete:
      operationId: deleteCandidate
      tags: [candidate]
      summary: Delete a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Candidate deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
      required: [status, message]
...
    AddCandidateInput:
      type: object
      properties:
        election_id:
          type: string
        name:
          type: string
        description:
          type: string
        email:
          type: string
        image_url:
          type: string
          nullable: true
      required: [election_id, name]
    Candidate:
      type: object
      properties:
        id:
          type: string
        election_id:
          type: string
        name:
          type: string
        description:
          type: string
        email:
          type: string
        image_url:
          type: string
          nullable: true
        votes:
          type: integer
      required: [id, election_id, name, description, email, votes]
    CandidatesResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Candidate"
      required: [status, message, data]
...